+++ config/urls.py
@@ -5,8 +5,8 @@ from django.urls import include, path
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('accounts/', include('django.contrib.auth.urls')),
     path('accounts/', include(('accounts.urls', 'accounts'), namespace='accounts')),
+    path('accounts/', include('django.contrib.auth.urls')),
     path('', include(('shop.urls', 'shop'), namespace='shop')),
 ]